Add trigger for log rotation and transfer
Periodically trigger log rotation and transfer. For each Logger task,
close current file, start writing to new file and upload newly closed
file.

It supports files with and without the ".log" suffix and with or without
automatic timestamp. If timestamp is not activated, it gives the new
file the old one's name with an incremented number at the end. If the
current file does not end with a number, it renames the current file by
adding a "1" at the end so that the new file ends with a "2". If
timestamp is activated, it just gives an updated timestamp to the new
file name.