include method to command file transfer from process server
Creates method send_file_transfer_command in which it is established
a client-server connection with a process server, referencing it's
name, and called a client method log_upload_file.

The send_file_transfer_command asks as arguments a name for the
process server, a temporary root CA object for  referencing user,
password and certificate (atributes in TmpRootCA class) for
authtentication purposes and a specific process server logfile name.